moving platform specific files toward libnyquist interface in AudioFileReader
