+++ src/internal/AudioFileReader.h
@@ -1,37 +1,10 @@
-/*
- * Copyright (C) 2010 Google Inc. All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions
- * are met:
- *
- * 1.  Redistributions of source code must retain the above copyright
- *     notice, this list of conditions and the following disclaimer.
- * 2.  Redistributions in binary form must reproduce the above copyright
- *     notice, this list of conditions and the following disclaimer in the
- *     documentation and/or other materials provided with the distribution.
- * 3.  Neither the name of Apple Computer, Inc. ("Apple") nor the names of
- *     its contributors may be used to endorse or promote products derived
- *     from this software without specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY APPLE AND ITS CONTRIBUTORS "AS IS" AND ANY
- * EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL APPLE OR ITS CONTRIBUTORS BE LIABLE FOR ANY
- * DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
- * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF
- * THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
+#ifndef AudioFileReaderMac_h
+#define AudioFileReaderMac_h
 
-#ifndef AudioFileReader_h
-#define AudioFileReader_h
+#include <memory>
 
-#include <stdlib.h>
-
-namespace WebCore {
+namespace WebCore
+{
 
 class AudioBus;
 
@@ -42,10 +15,29 @@ class AudioBus;
 std::unique_ptr<AudioBus> MakeBusFromMemory(const void * data, size_t dataSize, bool mixToMono, float sampleRate);
 
 std::unique_ptr<AudioBus> MakeBusFromFile(const char * filePath, bool mixToMono, float sampleRate);
-                                
-// May pass in 0.0 for sampleRate in which case it will use the AudioBus's sampleRate                               
+
+// May pass in 0.0 for sampleRate in which case it will use the AudioBus's sampleRate
 void writeBusToAudioFile(AudioBus* bus, const char* filePath, double fileSampleRate);
 
-} // namespace WebCore
+class AudioFileReader
+{
+    
+public:
+    
+    AudioFileReader(const char * filePath);
+    AudioFileReader(const void * data, size_t dataSize);
+    
+    ~AudioFileReader();
 
-#endif // AudioFileReader_h
+    // Returns nullptr if error
+    std::unique_ptr<AudioBus> createBus(float sampleRate, bool mixToMono);
+    
+private:
+    
+    void readInternal();
+
+};
+
+} // namespace WebCore
+ 
+#endif // AudioFileReaderMac_h
